requirements.txt
arabic-reshaper==3.0.0
asgiref==3.6.0
asn1crypto==1.5.1
Babel==2.12.1
Brotli==1.0.9
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cryptography==40.0.1
cssselect2==0.7.0
defusedxml==0.7.1
Deprecated==1.2.13
Django==4.1.7
django-allauth==0.53.1
django-environ==0.10.0
django-extensions==3.2.1
django-log-viewer==1.1.7
django-oauth-toolkit==2.2.0
django-phonenumber-field==7.0.2
django-pwa
djangorestframework==3.14.0
fonttools==4.39.3
html5lib==1.1
idna==3.4
jwcrypto==1.4.2
lxml==4.9.2
MarkupSafe==2.1.2
oauthlib==3.2.2
oscrypto==1.3.0
phonenumbers==8.13.8
Pillow==9.5.0
pycparser==2.21
pydyf==0.6.0
pyHanko==0.17.2
pyhanko-certvalidator==0.20.1
PyJWT==2.6.0
pyOpenSSL==23.1.0
pypdf==3.7.0
pyphen==0.14.0
pypng==0.20220715.0
python-bidi==0.4.2
python3-openid==3.2.0
pytz==2023.2
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.4.2
reportlab==3.6.12
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.3
svglib==1.5.1
tinycss2==1.2.1
typing-extensions==4.5.0
tzdata==2023.2
tzlocal==4.3
uritools==4.0.1
urllib3==1.26.15
weasyprint==58.1
webencodings==0.5.1
Werkzeug==2.2.3
wkhtmltopdf==0.2
wrapt==1.15.0
xhtml2pdf==0.2.9
zopfli==0.2.2
psycopg2-binary